% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callGenotype.R
\name{callGenotype}
\alias{callGenotype}
\title{Automatically call or flag alleles from NGS data}
\usage{
callGenotype(fb, tbase = NULL, motif = NULL)
}
\arguments{
\item{fb}{A `fishbone` object.
column names.}

\item{tbase}{A data.frame with thresholds which are locus specific. Thresholds are balance of alleles (B),
ignore threshold (IT), low run threshold (LRT) and relative stutter height (S).}

\item{motif}{A data.frame of loci motifs. Expected two columns with loci names (`locus`) and actual motifs (`motif`).}
}
\description{
Call or flag candidate alleles to construct a consensus genotype.
}
\details{
Function works on an individual run (one PCR reaction). Use R functionality to
apply this to sample * locus * run combination. The data should come from an NGS run as
processed by de Barba et al. (2016).

A note o finding stutters: For each candidate allele the sequence is shortened for the first
motif found given the locus.

De Barba, M., Miquel, C., Lobréaux, S., Quenette, P. Y., Swenson, J. E., & Taberlet, P. (2016).
High-throughput microsatellite genotyping in ecology: improved accuracy, efficiency, standardization
and success with low-quantity and degraded DNA. Molecular Ecology Resources, 1–16.
https://doi.org/10.1111/1755-0998.12594

Key for abbreviations:
A = allele
S = stutter
R = relative low threshold
L = low count threshold
D = disbalance

The result is appended two columns, one for called A and another for flagged.
Possible flags are:
* L = low amplification threshold
* N = no stutter
* D = alleles not in balance
* M = multiple alleles

Algorithm is as follows:
- descendingly sort A according to number of reads
- find number of reads for highest A -> maxA

0. find max allele height
1. check that at least max A is above L threshold
1a. if not, flag all alleles with "L"
2. see if allele has stutter
2a. if yes, mark as called
2aa. if A in disbalance (A < D), flag as "D"
2ab. mark stutter as such $stutter = TRUE
2b. if no, check AlleleWithNoStutterHeight
2ba. if x > AlleleWithNoStutterHeight, add flag "N"
2bb. if x < AlleleWithNoStutterHeight, ignore allele
3. if number of unflagged alleles is more than 2, add flag "M" to all

Output should be all alleles and their stutters.
}
