% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbone-package.R
\docType{package}
\name{fishbone-package}
\alias{fishbone-package}
\alias{fishbone}
\title{Automatic allele calling or flagging.}
\description{
A package for automatic allele calling or flagging.
}
\details{
\tabular{ll}{
Package: \tab fishbone\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2017-07-08\cr
License: \tab GPL-2\cr
LazyData: \tab no\cr
}
}
\author{
Roman Lustrik (\email{roman.lustrik@biolitika.si})

Maintainer: Roman Lustrik (\email{roman.lustrik@biolitika.si})
}
\keyword{hplot,}
\keyword{manip}
