% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mt}
\alias{mt}
\title{Lookup table of calling parameters}
\format{A data.table.}
\source{
LIFE DinAlpBear project
}
\usage{
mt
}
\description{
Algorithm of calling alleles uses some prio information about alleles and stutters.
This table provide these values. Tweak or create according to your needs.
}
\author{
Roman Lustrik, Tomaz Skrbinsek, Maja Jelencic, Meta Konec
}
\keyword{dataset}
