% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fishbone}
\alias{plot.fishbone}
\title{Plot fishbone object.}
\usage{
\method{plot}{fishbone}(x, ...)
}
\arguments{
\item{x}{\code{fishbone} object.}

\item{...}{Currently not used.}
}
\description{
Plots candidate alleles and colors those that are not expected given the
nucleotide repeat length. For instance, if a difference in length between two
alleles is not say 4 for tetranucleotides, it could be due to tag jump and the
allele is probably not what you're after.
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}
