% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotRuns}
\alias{plotRuns}
\title{Plot a list of \code{fishbone} objects.}
\usage{
plotRuns(x)
}
\arguments{
\item{x}{A list of \code{fishbone} objects.}
}
\description{
Plot a list of \code{fishbone} objects.
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}
