% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fishbone}
\alias{print.fishbone}
\title{Improves printing of \code{fishbone} objects.}
\usage{
\method{print}{fishbone}(x, ...)
}
\arguments{
\item{x}{Object of class \code{fishbone}.}

\item{...}{Currently not used.}
}
\description{
To prevent cluttering of the output, this is a convencience functionality which
truncates Sequence to a few characters. Coerce to \code{data.frame} for the original
values.
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}
