% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{smp1}
\alias{smp1}
\title{Reads of microsatellite sequences of a non-invasive sample}
\format{A data.table.}
\source{
LIFE DinAlpBear project
}
\usage{
smp1
}
\description{
NGS (next generation sequencer) data from a project contains sequences, their read count,
sample name, repeat number (PCR plate), marker (locus) designation, run name (arbitrary), position
(on PCR plate), length of sequence, combination of tags used and of course, the microsatellite sequence
itself.
}
\author{
Roman Lustrik, Tomaz Skrbinsek, Maja Jelencic, Meta Konec
}
\keyword{dataset}
