% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.fishbone}
\alias{as.fishbone}
\title{Coerces a data.frame (like object) to a `fishbone` object.}
\usage{
as.fishbone(x, motif.length)
}
\arguments{
\item{x}{Object of class \code{data.frame} or similar.}

\item{motif.length}{Integer. Length of the repeating motif, locus specific.}
}
\description{
Expected columns are Sample_Name, Plate, Marker, Read_Count, Allele, Sequence.
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}
