% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callGenotype.R
\name{callGenotype}
\alias{callGenotype}
\title{Automatically call or flag alleles from NGS data}
\usage{
callGenotype(fb, tbase = NULL, motif = NULL)
}
\arguments{
\item{fb}{A `fishbone` object.
column names.}

\item{tbase}{A data.frame with thresholds which are locus specific. Thresholds are balance of alleles (B),
ignore threshold (IT), low run threshold (LRT) and relative stutter height (S).}

\item{motif}{A data.frame of loci motifs. Expected two columns with loci names (`locus`) and actual motifs (`motif`).}
}
\description{
Call or flag candidate alleles to construct a consensus genotype.
}
\details{
Function works on an individual run (one PCR reaction). Use R functionality to
apply this to sample * locus * run combination. The data should come from an NGS run as
processed by de Barba et al. (2016).

A note o finding stutters: For each candidate allele the sequence is shortened for the first
motif found given the locus.

De Barba, M., Miquel, C., Lobréaux, S., Quenette, P. Y., Swenson, J. E., & Taberlet, P. (2016).
High-throughput microsatellite genotyping in ecology: improved accuracy, efficiency, standardization
and success with low-quantity and degraded DNA. Molecular Ecology Resources, 1–16.
https://doi.org/10.1111/1755-0998.12594

Key for abbreviations:
A = allele
S = stutter
R = relative low threshold
LRT = low run threshold
B = balance

The result is appended two columns, one for called A and another for flagged.
Possible flags are:
* L = low run threshold
* N = no stutter
* B = alleles not in balance
* M = multiple alleles
* R = relative contribution of allele(s) to run

Algorithm is as follows:
- descendingly sort A according to number of reads
- find number of reads for highest A -> maxA

for each A:
compare everything according to the highest read count, calculate relative size to maxA -> rs
1. if signal is below relative low threshold (R), remove this allele completely
2. if below low run threshold (LRT), add flag "L" to all alleles
3. if stutter for A found (relative height correct), call, otherwise flag as "N"
3b. also return stutter
4. if disbalance in alleles (rs <= B), add flag "B" to low A
5. if number of unflagged A > 2, add flag "M" to all alleles

Output should be all alleles and their stutters.
}
