% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callGenotype.R
\name{fetchTH}
\alias{fetchTH}
\title{Fetch threshold values from some database.}
\usage{
fetchTH(x, stat, locus)
}
\arguments{
\item{x}{A slice of `fishbone` object.}

\item{stat}{Which statistic to fetch, e.g. "IT", "B", "S", "LRT".}

\item{locus}{Character. For which locus to fetch statistics?}
}
\description{
Fetch threshold values from some database.
}
\author{
Roman Lustrik (roman.lustri@biolitika.si)
}
