% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findStutter.R
\name{findStutter}
\alias{findStutter}
\title{Find locus specific stutter.}
\usage{
findStutter(x, fb, motif)
}
\arguments{
\item{x}{A `fishbone` object.}

\item{fb}{A `fishbone` object with cancidate stutter alleles.}
}
\value{
A character string of the row name of the stutter(s). If stutter is not found, it
returns NULL.
}
\description{
From a `fishbone` object, try to find a stutter.
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}
